added query functionality
