+++ src/components/groups/custom/activities/ActivityParticipants.tsx
@@ -6,6 +6,7 @@ import { TimestampToDateString } from "@/src/utils/getCurrentDate";
 import Link from "next/link";
 import React from "react";
 import { SuspenseGroupActivityFetchType } from "./GroupActivityData";
+import ActivityParticipantsList from "./ActivityParticipantsList";
 
 export default async function ActivityParticipants({
   groupID,
@@ -28,29 +29,7 @@ export default async function ActivityParticipants({
 
     return (
       <DefaultCard className="w-full flex flex-col items-start justify-center gap-2">
-        <h1 className="text-custom-dark-text font-semibold">
-          Participants ( {Object.keys(participantsData).length} )
-        </h1>
-        <InnerContainer className="w-full">
-          {Object.keys(participantsData).map((memberID: string) => {
-            const date = participantsData[memberID].dateJoined;
-            const dateStr = TimestampToDateString(date);
-            return (
-              <Link
-                key={memberID}
-                href={`/members/${memberID}`}
-                className="w-full flex flex-col items-start justify-center py-2 px-3 duration-200 hover:bg-custom-light-text"
-              >
-                <h1 className="text-custom-dark-text font-semibold">
-                  {memberID}
-                </h1>
-                <h4 className="text-custom-grey-text text-sm">
-                  Participated on: {dateStr}
-                </h4>
-              </Link>
-            );
-          })}
-        </InnerContainer>
+        <ActivityParticipantsList participantsData={participantsData} />
       </DefaultCard>
     );
   } catch (err: any) {
